# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Star Mall service.",
        "title": "Star Mall API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/user/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除当前登录用户的账户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "description": "Delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/forgetPassword": {
            "patch": {
                "description": "用户忘记密码, 重置密码接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "忘记密码",
                "parameters": [
                    {
                        "description": "Forget password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ForgetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "password updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/get": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户信息",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "用户通过 email 或 userID 登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "put": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册接口",
                "parameters": [
                    {
                        "description": "Register request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/update": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "Update user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/user/update/password": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "Update password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "password updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.User": {
            "description": "用户模型",
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "deleteTime": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "imageID": {
                    "description": "uid",
                    "type": "integer"
                },
                "isVip": {
                    "type": "boolean"
                },
                "lastIP": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "规则ID, 用于分别用户类型",
                    "type": "integer"
                },
                "sex": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
...
        "handler.DeleteRequest": {
            "type": "object",
            "required": [
                "email",
                "verifyCode"
            ],
            "properties": {
                "email": {
                    "description": "@Description 邮箱",
                    "type": "string"
                },
                "verifyCode": {
                    "description": "@Description 验证码",
                    "type": "string"
                }
            }
        },
        "handler.ForgetPasswordRequest": {
            "type": "object",
            "required": [
                "email",
                "newPassword",
                "verifyCode"
            ],
            "properties": {
                "email": {
                    "description": "@Description 邮箱\n@Required true\n@Example \"\u003cEMAIL\u003e\"",
                    "type": "string"
                },
                "newPassword": {
                    "description": "@Description 新密码\n@Required true\n@Example \"\u003cPASSWORD\u003e\"",
                    "type": "string",
                    "minLength": 6
                },
                "verifyCode": {
                    "description": "@Description 验证码\n@Required true\n@Example \"123456\"",
                    "type": "string"
                }
            }
        },
...
        "handler.LoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "handler.LoginResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "email",
                "newPassword",
                "oldPassword",
                "verifyCode"
            ],
            "properties": {
                "email": {
                    "description": "@Description 邮箱\n@Required true\n@Example \"\u003cEMAIL\u003e\"",
                    "type": "string"
                },
                "newPassword": {
                    "description": "@Description 新密码\n@Required true\n@Example \"\u003cPASSWORD\u003e\"",
                    "type": "string",
                    "minLength": 6
                },
                "oldPassword": {
                    "description": "@Description 原密码\n@Required true\n@Example \"\u003cPASSWORD\u003e\"",
                    "type": "string",
                    "minLength": 6
                },
                "verifyCode": {
                    "description": "@Description 验证码\n@Required true\n@Example \"12AB56\"",
                    "type": "string"
                }
            }
        },
        "handler.UpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "@Description 邮箱\n@Required false\n@Example \"example@mail.com\"",
                    "type": "string"
                },
                "name": {
                    "description": "@Description 用户名\n@Required false\n@Example \"张三\"",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "phone": {
                    "description": "@Description 手机号\n@Required false\n@Example \"+8613800138000\"",
                    "type": "string"
                },
                "sex": {
                    "description": "@Description 性别(203:未知 201:男 202:女)\n@Required false\n@Enum 201 202 203\n@Example 1",
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                }
            }
        },
...
        "handler.UserRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone",
                "sex"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                },
                "verificationCode": {
                    "type": "string"
                }
            }
        },
        "utils.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "输入 JWT token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}